Other mouse button and event suppression (commented out).
This increases the result size to 1115 bytes. However, it does add
middle and right mouse button support and suppresses mouse and
keyboard default actions so that the right click menu doesn't appear
and so that tab can be used.

Leave it commented out since this goes beyond the limit but it is
interesting nonetheless.